feat(gateway): resolve GitHub capsule sources in the uplink before install
POST /api/capsules forwarded its source verbatim to the kernel's
path-only InstallCapsule handler, which rejects remote-shaped sources,
so a gateway-only integrator could not install a GitHub capsule the way
the CLI can. The gateway, an uplink like the CLI, now resolves
GitHub-shaped sources itself: fetch the repo's latest release, select
the matching .capsule asset (lone, or named via a new optional capsule
field), stream it to a temp file with a 50 MB cap, then call the
existing InstallCapsule with that LOCAL path. The kernel keeps rejecting
URLs; the fetch lives only in the gateway.

The five pure GitHub source-parsing/asset-selection helpers move into
astrid-capsule-install::github_source, a network-free module shared by
the CLI and gateway. astrid-capsule-install and astrid-kernel gain no
network dependency.

Closes #960
