+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 
 - **`astrid agent create --clone <source>` replicates an existing agent's full profile and state in one atomic admin call.** `--inherit-from` copies state only (env/KV/secrets) and deliberately leaves the capability profile empty, so there was no first-class way to *replicate* an agent — duplicating one meant rebuilding its groups, grants, revokes, egress, process allow-list, and quotas by hand via `caps grant` / `agent modify` / `quota set`. `AgentCreate` gains `clone_from: Option<PrincipalId>` and `allow_admin_clone: bool` (both `#[serde(default)]`, so older serialized requests deserialize unchanged). When `clone_from` is set, the kernel builds the new principal as a full replica of the source — groups + grants + revokes + network egress + process-spawn allow-list + quotas — and then performs the same env/KV/secret state copy `inherit_from` does. Deliberately **NOT** copied: the source's `auth` (each principal keeps its own keys/authenticators — cloning is profile+state, never credentials) and its `enabled` flag (a fresh clone is always enabled, even if the source was disabled). `--clone` is mutually exclusive with the profile/quota-shaping flags (`--group`, `--egress`, `--process-allow`, `--inherit-from`, `--memory/--timeout/--storage/--processes`): the CLI enforces this via clap `conflicts_with` and the kernel rejects a request that sets both — defense in depth against a hand-built request. **Admin-source guard:** cloning a source that resolves to the universal `*` (the built-in `admin` group, a `*` grant, or a custom `unsafe_admin` group — resolved through the live `GroupConfig`, not a literal scan) is rejected unless the operator passes `--unsafe-admin`, mirroring the acknowledgement on `caps grant '*'` and `group create --caps '*'` — so `--clone default` cannot silently mint a second admin. The source is validated to exist and not be the new principal itself before any write runs. The HTTP management API constructs `clone_from: None` (an API-level clone is a follow-up). Closes #927.
 
+- **Gateway `POST /api/capsules` now resolves GitHub sources (`@org/repo`, `github.com/org/repo`, `https://github.com/org/repo`) to a local archive in the uplink before install; the daemon still never fetches URLs.** Previously the gateway forwarded `source` verbatim to the kernel's path-only `InstallCapsule` handler, which rejects every remote-shaped source — so an integrator that only speaks gateway HTTP could not install a registry/GitHub capsule even though the CLI could (the CLI resolves GitHub sources client-side and hands the kernel a local path). The gateway, an uplink like the CLI, now does the same: for a GitHub-shaped source it fetches the repo's latest release, selects the matching `.capsule` asset (the lone one, or the one named via a new optional `capsule` request field for a multi-capsule release), streams it to a temp file with a hard 50 MB cap, and calls the existing `InstallCapsule` with that LOCAL path (`workspace` forced `false`). The kernel install path is unchanged and still rejects URLs — the fetch lives ONLY in the gateway. The five pure GitHub source-parsing/asset-selection helpers move into `astrid-capsule-install::github_source` (a network-free module the CLI and gateway share); the kernel and the install crate gain no network dependency. The gateway implements a narrower path than the CLI on purpose: no clone-and-build fallback, no local-Cargo auto-build, no multi-asset "install all" — it requires resolving to exactly one `.capsule`. New 404 (no release / no `.capsule` asset) and 400 (ambiguous multi-capsule release, or archive over 50 MB) responses. Closes #960.
+
 - **The `astrid mcp serve` shim now elicits interactive user consent when the sage-mcp broker reports an untrusted ingress, replacing the static operator allow-list.** Previously the broker gated its state-mutating `tools/call` front door on an operator-maintained `trusted_ingress_ids` list; an ingress that was not pre-pinned was hard-denied. The broker now replies an `ingress_approval_required` signal for a not-yet-trusted ingress, and the shim turns it into a user consent prompt via the MCP elicitation capability ("Allow Astrid tool calls from this session?"). On accept the shim forwards `astrid.v1.request.mcp.ingress.respond` `{ req_id, accept }`, the broker records trust keyed on the kernel-stamped caller `source_id` (never a body field), and the shim re-sends the original call. Fail-secure: a client without elicitation, a decline / cancel, an elicit error, or a broker that could not persist the grant all return an MCP error to the client and record no trust. Manual end-to-end verification of the live elicit flow is pending (needs an elicitation-capable client, Claude Code ≥2.1.76). Paired with the sage-mcp broker change. Closes #917.
 
 ### Removed
